feat(zink): move zink to the top level with examples
