--- examples/.cargo/config.toml
@@ -1,7 +1,0 @@
-[build]
-target = "wasm32-unknown-unknown"
-
-[target.wasm32-unknown-unknown]
-rustflags = [
-    "-C", "link-args=--import-memory",
-]
