--- examples/Cargo.toml
@@ -1,18 +1,0 @@
-[workspace.package]
-edition = "2021"
-version = "0.0.0"
-authors = [ "clearloop" ]
-license = "GPL-v3.0"
-
-[workspace]
-resolver = "2"
-members = [
-  "addition",
-  "fibonacci",
-  "if-else",
-  "log",
-  "storage"
-]
-
-[workspace.dependencies]
-zink = { path = "../zink" }
