+++ examples/addition.rs
@@ -1,8 +1,6 @@
 //! Addition example.
-#![no_std]
+#![cfg_attr(target_arch = "wasm32", no_std)]
 
-// for the panic handler.
-#[cfg(not(test))]
 extern crate zink;
 
 /// Adds two numbers together.
@@ -10,3 +8,7 @@ extern crate zink;
 pub fn addition(x: u64, y: u64) -> u64 {
     x + y
 }
+
+// This is required when because we want to
+// build this with cargo b.
+fn main() {}
