--- examples/addition/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "addition"
-version = "0.1.0"
-edition.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-zink.workspace = true
