+++ examples/if-else.rs
@@ -1,8 +1,6 @@
 //! if-else example.
-#![no_std]
+#![cfg_attr(target_arch = "wasm32", no_std)]
 
-// for the panic handler.
-#[cfg(not(test))]
 extern crate zink;
 
 /// Simple if-else condition
@@ -14,3 +12,5 @@ pub extern "C" fn if_else(x: u64, y: u64) -> u64 {
         y
     }
 }
+
+fn main() {}
