--- examples/if-else/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "if-else"
-version = "0.1.0"
-edition.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-zink.workspace = true
