--- examples/storage/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "storage"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-zink.workspace = true
