--- zink/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "zink"
-description = "Standard library for zink projects."
-documentation = "https://docs.rs/zink"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-zink-codegen.workspace = true
