--- zink/README.md
@@ -1,7 +1,0 @@
-# Zink
-
-The standard library for zink projects.
-
-## LICENSE
-
-GPL-3.0
